Move insert_release/insert_release_sc helpers to a more reusable location
